[BE] MOVE: Search 관련 폴더 생성 및 이동 #205
