+++ src/backend/main-server/main/src/main/java/com/kickzo/main/search/document/UserDocument.java
@@ -1,22 +1,30 @@
-package com.kickzo.main;
+package com.kickzo.main.search.document;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.Document;
 import org.springframework.data.elasticsearch.annotations.Field;
 import org.springframework.data.elasticsearch.annotations.FieldType;
+
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
 @Document(indexName = "search")
-public class RoomDocument {
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserDocument {
 	@Id
-	private Long roomId;
+	private Long userId;
 
 	@Field(type = FieldType.Text)
-	private String title;
+	private String nickname;
 
-	@Field(type = FieldType.Boolean)
-	private boolean isPublic;
+	@Field(type = FieldType.Text)
+	private String stateMessage;
+
+	@Field(type = FieldType.Text)
+	private String profileImageUrl;
 }
