+++ src/backend/main-server/main/src/main/java/com/kickzo/main/search/repository/RoomSearchRepository.java
@@ -1,8 +1,10 @@
-package com.kickzo.main;
+package com.kickzo.main.search.repository;
 
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
+
+import com.kickzo.main.search.document.RoomDocument;
 
 @Repository
 public interface RoomSearchRepository extends ElasticsearchRepository<RoomDocument, Long> {
