+++ src/backend/main-server/main/src/main/java/com/kickzo/main/search/repository/UserSearchRepository.java
@@ -1,9 +1,11 @@
-package com.kickzo.main;
+package com.kickzo.main.search.repository;
 
 import java.util.List;
 
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 import org.springframework.stereotype.Repository;
+
+import com.kickzo.main.search.document.UserDocument;
 
 @Repository
 public interface UserSearchRepository extends ElasticsearchRepository<UserDocument, Long> {
